app/requirements.txt
backcall==0.2.0
Brotli==1.0.9
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dash==1.8.0
dash-building-blocks==0.1.2
dash-core-components==1.7.0
dash-html-components==1.0.2
dash-renderer==1.2.3
dash-table==4.6.0
dash-table-experiments==0.6.0
decorator==4.4.2
feather-format==0.4.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
kiwisolver==1.3.1
line-profiler==3.1.0
MarkupSafe==1.1.1
matplotlib==3.3.3
# mkl-fft==1.2.0
# mkl-random==1.1.1
# mkl-service==2.3.0
# more-itertools==8.6.0
numpy #@ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
pandas #@ file:///tmp/build/80754af9/pandas_1602088120436/work
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyarrow==2.0.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz # @ file:///tmp/build/80754af9/pytz_1606604771399/work
Quandl==3.5.3
requests==2.25.0
retrying==1.3.3
six # @ file:///tmp/build/80754af9/six_1605205327372/work
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
